requirements.txt
paho-mqtt==1.5.0
PyYAML==5.1.2
six==1.12.0
websocket-client==0.56.0
